+++ backend/src/main/java/com/woowahanbank/backend/domain/user/repository/UserRepository.java
@@ -3,17 +3,13 @@ package com.woowahanbank.backend.domain.user.repository;
 import java.util.List;
 import java.util.Optional;
 
-import javax.swing.text.html.Option;
-
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 import com.woowahanbank.backend.domain.user.domain.User;
 
-import io.lettuce.core.dynamic.annotation.Param;
-
 @Repository
-public interface UserRepository extends JpaRepository<User, Long> {
+public interface UserRepository extends JpaRepository<User, Integer> {
 	Optional<User> findByUserId(String userId);
 
 	Optional<User> findByNickname(String nickname);
